Move tests to unit test directory, switch travis to using grunt for the test runner
