+++ test/unit/test-commands.js
