+++ test/unit/test-fetch.js
@@ -2,6 +2,7 @@
 
 const setup = require('./include/setup');
 const sinon = require('sinon');
+const FetchSummary = require('../../src/responses/FetchSummary');
 
 var git, sandbox;
 
@@ -25,7 +26,7 @@ exports.push = {
 
    'git generates a fetch summary': function (test) {
       git.fetch('r', 'b', function (err, result) {
-         test.ok(result instanceof require('../src/responses/FetchSummary'));
+         test.ok(result instanceof FetchSummary);
          test.done();
       });
       setup.closeWith('');
@@ -64,24 +65,21 @@ exports.push = {
    },
 
    'parses new tags': function (test) {
-      var parser = require('../src/responses/FetchSummary');
-      var summary = parser.parse(' * [new tag]         0.11.0     -> 0.11.0');
+      var summary = FetchSummary.parse(' * [new tag]         0.11.0     -> 0.11.0');
 
       test.same(summary.tags, [{ name: '0.11.0', tracking: '0.11.0' }]);
       test.done();
    },
 
    'parses new branches': function (test) {
-      var parser = require('../src/responses/FetchSummary');
-      var summary = parser.parse(' * [new branch]         master     -> origin/master');
+      var summary = FetchSummary.parse(' * [new branch]         master     -> origin/master');
 
       test.same(summary.branches, [{ name: 'master', tracking: 'origin/master' }]);
       test.done();
    },
 
    'parses remote': function (test) {
-      var parser = require('../src/responses/FetchSummary');
-      var summary = parser.parse('From https://github.com/steveukx/git-js');
+      var summary = FetchSummary.parse('From https://github.com/steveukx/git-js');
 
       test.same(summary.remote, 'https://github.com/steveukx/git-js');
       test.done();
