+++ test/unit/test-log.js
