+++ test/unit/test-stash.js
